Add more functionality to build-list
The lambda that build-list takes can now optionally take a second
function argument that will add a whole list to the built list.
This makes for a faster concat function.

I also added a flatten function, which is like concat, but takes
a single list and flattens it instead of taking multiple args.